+++ packages/permissionless/accounts/nexus/toNexusSmartAccount.ts
@@ -1,19 +1,17 @@
-import type {
-    Account,
-    Assign,
-    Chain,
-    OneOf,
-    Prettify,
-    Transport,
-    TypedDataDefinition,
-    WalletClient
-} from "viem"
 import {
+    type Account,
     type Address,
+    type Assign,
+    type Chain,
     type Client,
     type Hex,
     type JsonRpcAccount,
     type LocalAccount,
+    type OneOf,
+    type Prettify,
+    type Transport,
+    type TypedDataDefinition,
+    type WalletClient,
     concat,
     concatHex,
     domainSeparator,