+++ packages/permissionless/actions/erc7579/uninstallModules.ts
@@ -4,8 +4,7 @@ import {
     type SmartAccount,
     sendUserOperation
 } from "viem/account-abstraction"
-import { getAction } from "viem/utils"
-import { parseAccount } from "viem/utils"
+import { getAction, parseAccount } from "viem/utils"
 import { AccountNotFoundError } from "../../errors/index.js"
 import {
     type EncodeUninstallModuleParameters,