+++ packages/permissionless/actions/etherspot.ts
@@ -1,6 +1,4 @@
-import {
+export {
     type GetGasPriceResponseReturnType,
     getUserOperationGasPrice
 } from "./etherspot/getUserOperationGasPrice.js"
-
-export { type GetGasPriceResponseReturnType, getUserOperationGasPrice }