+++ packages/permissionless/actions/passkeyServer.ts
@@ -1,29 +1,17 @@
-import {
+export {
     type StartRegistrationParameters,
     type StartRegistrationReturnType,
     startRegistration
 } from "./passkeyServer/startRegistration.js"
 
-import {
+export {
     type VerifyRegistrationParameters,
     type VerifyRegistrationReturnType,
     verifyRegistration
 } from "./passkeyServer/verifyRegistration.js"
 
-import {
+export {
     type GetCredentialsParameters,
     type GetCredentialsReturnType,
     getCredentials
 } from "./passkeyServer/getCredentials.js"
-
-export {
-    type StartRegistrationParameters,
-    type StartRegistrationReturnType,
-    type VerifyRegistrationParameters,
-    type VerifyRegistrationReturnType,
-    type GetCredentialsParameters,
-    type GetCredentialsReturnType,
-    startRegistration,
-    verifyRegistration,
-    getCredentials
-}