+++ packages/permissionless/actions/smartAccount.ts
@@ -1,9 +1,7 @@
-import { sendTransaction } from "./smartAccount/sendTransaction.js"
+export { sendTransaction } from "./smartAccount/sendTransaction.js"
 
-import { signMessage } from "./smartAccount/signMessage.js"
+export { signMessage } from "./smartAccount/signMessage.js"
 
-import { signTypedData } from "./smartAccount/signTypedData.js"
+export { signTypedData } from "./smartAccount/signTypedData.js"
 
-import { writeContract } from "./smartAccount/writeContract.js"
-
-export { sendTransaction, signMessage, signTypedData, writeContract }
+export { writeContract } from "./smartAccount/writeContract.js"