+++ packages/permissionless/experimental/pimlico/index.ts
@@ -1,2 +1,1 @@
-import { prepareUserOperationForErc20Paymaster } from "./utils/prepareUserOperationForErc20Paymaster.js"
-export { prepareUserOperationForErc20Paymaster }
+export { prepareUserOperationForErc20Paymaster } from "./utils/prepareUserOperationForErc20Paymaster.js"